Update pipboy.cc for full pagination
Full implementation of pagination for the pipboy. Status, Data, and Automaps, and their subpages can now paginate when entries exceed a range which fits in the page.

The 'Video Archives' page was not given pagination, as it seem unlikely it would be needed, and there currently is a limit to number of videos.